Refactoring follow review
